+++ .gitignore
@@ -36,7 +36,7 @@ src/output/
 #  before PyInstaller builds the exe, so as to inject date/other infos into it.
 *.manifest
 *.spec
-
+notebooks/*.pkl
 # Installer logs
 pip-log.txt
 pip-delete-this-directory.txt
