--- src/models/ner_model_1/meta.json
@@ -1,34 +1,0 @@
-{
-  "lang":"en",
-  "name":"pipeline",
-  "version":"0.0.0",
-  "spacy_version":">=3.8.4,<3.9.0",
-  "description":"",
-  "author":"",
-  "email":"",
-  "url":"",
-  "license":"",
-  "spacy_git_version":"85cc763",
-  "vectors":{
-    "width":0,
-    "vectors":0,
-    "keys":0,
-    "name":null,
-    "mode":"default"
-  },
-  "labels":{
-    "ner":[
-      "INGREDIENT",
-      "QUANTITY"
-    ]
-  },
-  "pipeline":[
-    "ner"
-  ],
-  "components":[
-    "ner"
-  ],
-  "disabled":[
-
-  ]
-}
