--- src/models/ner_model_1/vocab/vectors.cfg
@@ -1,3 +1,0 @@
-{
-  "mode":"default"
-}
