--- src/models/ner_model_2/ner/moves
@@ -1,1 +1,0 @@
-Ç•movesŸ®{"0":{},"1":{"INGREDIENT":-1,"QUANTITY":-2},"2":{"INGREDIENT":-1,"QUANTITY":-2},"3":{"INGREDIENT":-1,"QUANTITY":-2},"4":{"":1,"INGREDIENT":-1,"QUANTITY":-2},"5":{"":1}}£cfgÅßneg_key¿
