--- src/models/ner_model_2/vocab/vectors.cfg
@@ -1,3 +1,0 @@
-{
-  "mode":"default"
-}
