--- src/models/ner_model_3/vocab/vectors.cfg
@@ -1,3 +1,0 @@
-{
-  "mode":"default"
-}
